Add get and set navigationDefinition on Chrome Service
