+++ packages/core/chrome/core-chrome-browser-internal/src/chrome_service.tsx
@@ -31,6 +31,8 @@ import type {
   ChromeStyle,
   ChromeProjectNavigation,
   ChromeSetProjectBreadcrumbsParams,
+  NavigationTreeDefinition,
+  AppDeepLinkId,
 } from '@kbn/core-chrome-browser';
 import type { CustomBrandingStart } from '@kbn/core-custom-branding-browser';
 import type {
@@ -272,6 +274,15 @@ export class ChromeService {
       validateChromeStyle();
       projectNavigation.setProjectNavigation(config);
     };
+
+    function setProjectNavigationDefinition<
+      LinkId extends AppDeepLinkId = AppDeepLinkId,
+      Id extends string = string,
+      ChildrenId extends string = Id
+    >(navigationTree$: Observable<NavigationTreeDefinition<LinkId, Id, ChildrenId>>) {
+      validateChromeStyle();
+      projectNavigation.setNavigationDefinition(navigationTree$);
+    }
 
     const setProjectBreadcrumbs = (
       breadcrumbs: ChromeBreadcrumb[] | ChromeBreadcrumb,
@@ -526,6 +537,8 @@ export class ChromeService {
         setProjectUrl,
         setProjectName,
         setNavigation: setProjectNavigation,
+        setNavigationDefinition: setProjectNavigationDefinition,
+        getNavigationDefinition$: () => projectNavigation.getNavigationDefinition$(),
         setSideNavComponent,
         setBreadcrumbs: setProjectBreadcrumbs,
         getActiveNavigationNodes$: () => projectNavigation.getActiveNodes$(),
